PR_code/requirements.txt
tensorflow
torch
transformers
datasets
matplotlib
seaborn
pandas
scikit-learn
nltk
rouge-score
PyYAML
wordcloud
